requirements.txt
langchain
python-dotenv
langchain_community
gitpython
faiss-cpu
langchain-openai
pydantic
lizard
javalang